Introduce mock interfaces module
