+++ tests/cli/functional/conftest.py
@@ -23,11 +23,11 @@ from nucypher.blockchain.eth.agents import ContractAgency
 from nucypher.blockchain.eth.interfaces import BlockchainInterface, BlockchainInterfaceFactory
 from nucypher.blockchain.eth.registry import InMemoryContractRegistry
 from nucypher.config.characters import UrsulaConfiguration
-from nucypher.utilities.sandbox.blockchain import MockBlockchain
 from nucypher.utilities.sandbox.constants import MOCK_PROVIDER_URI
 from tests.cli.functional.test_ursula_local_keystore_cli_functionality import NUMBER_OF_MOCK_ACCOUNTS, \
     KEYFILE_NAME_TEMPLATE
-from tests.fixtures import _make_testerchain, make_token_economics, make_test_registry_source_manager
+from tests.fixtures import _make_testerchain, make_token_economics
+from tests.mock.interfaces import make_mock_registry_source_manager, MockBlockchain
 from tests.mock.agents import FAKE_RECEIPT, MockContractAgency
 
 
@@ -59,7 +59,7 @@ def test_registry():
 
 @pytest.fixture(scope='module')
 def test_registry_source_manager(mock_testerchain, test_registry):
-    return make_test_registry_source_manager(blockchain=mock_testerchain, test_registry=test_registry)
+    return make_mock_registry_source_manager(blockchain=mock_testerchain, test_registry=test_registry)
 
 
 @pytest.fixture(scope='module', autouse=True)
